refactored interfaces, issue with reader
