--- handlers/ibuffer.go
@@ -1,7 +1,0 @@
-package handlers
-
-import "github.com/smikelson75/parser/readers"
-
-type IBuffer interface {
-	Get(pattern readers.TokenPattern) (*string, error)
-}
