+++ handlers/interfaces/IPattern.go
@@ -1,4 +1,4 @@
-package patterns
+package interfaces
 
 type IPattern interface {
 	Pattern() string
