+++ handlers/interfaces/IPatternsBuffer.go
@@ -1,0 +1,5 @@
+package interfaces
+
+type IPatternsBuffer interface {
+	GetTo(start, end IPattern) (*string, error)
+}
