+++ handlers/interfaces/ibuffer.go
@@ -1,0 +1,5 @@
+package interfaces
+
+type IBuffer interface {
+	Get(pattern IPattern) (*string, error)
+}
