+++ handlers/interfaces/ireader.go
@@ -1,0 +1,6 @@
+package interfaces
+
+type IReader interface {
+	Get(pattern IPattern) (*string, error)
+	GetTo(start, end IPattern) (*string, error)
+}
